+++ CHANGELOG.md
@@ -5,7 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-(Please put changes here.)
+- Fix `signature::canonical_uri` to produce invalid uri when `Region::Custom::endpoint` has `/` as path (eg. `localhost:5555/`).
 
 ## [0.48.0] - 2022-04-24